Fix for losing expiry date/time as provided by auth
Pass on a session expiry date/time as provided by the authentication handler
via the produced UserInfo object. The current code carries the expiry date/time
along until the moment of the redirect after logging in. At that point the info
is lost and the expiry date/time is not used eventually for setting the
certificate expiry.